+++ common/src/types/types.ts
@@ -1,0 +1,11 @@
+export enum SymbolDataResolution {
+  _1_MINUTE = '1minute',
+  _5_MINUTES = '5minutes',
+  _15_MINUTES = '15minutes',
+  _30_MINUTES = '30minutes',
+  HOUR = 'hour',
+  _4_HOURS = '4hours',
+  DAY = 'day',
+  WEEK = 'week',
+  MONTH = 'month',
+}