+++ frontend/src/components/AssetPrice.vue
@@ -1,10 +1,7 @@
 <template>
-  <div v-if="price" class="flex flex-wrap overflow-hidden p-6 mb-6 bg-white rounded-lg">
+  <div v-if="price" class="flex flex-wrap overflow-hidden gap-2">
     <div class="text-5xl font-bold">
-      {{ formatCurrency(price.price, symbol.currency, false) }}
-    </div>
-    <div class="text-gray-400 font-bold">
-      <!-- currency -->
+      {{ formatCurrency(price.price, symbol.currency, false, 'en-us', 'never') }}
     </div>
     <PriceChange :compare-from="price.previousClose" :compare-to="price.price" />
     <PriceChange
@@ -14,31 +11,31 @@
       type="total"
     />
   </div>
+  <div v-else>
+    {{ $t('details.no_price') }}
+  </div>
 </template>
 
 <script setup lang="ts">
 import { AssetPrice, AssetSymbol } from '@common/models';
 import PriceChange from '@components/PriceChange.vue';
 import { CurrencyUtils } from '@goplan-finance/utils';
-import { onBeforeMount, onUnmounted, ref } from 'vue';
+import { onUnmounted, ref } from 'vue';
 
 const props = defineProps<{
   symbol: AssetSymbol;
 }>();
 
-let liveSubscription = null;
+const liveSubscription = ref<Parse.LiveQuerySubscription>();
+const price = ref<AssetPrice>();
 
-const price = ref(null);
-
-onBeforeMount(async () => {
-  liveSubscription = await AssetPrice.liveQuery(props.symbol, assetPrice => {
-    price.value = assetPrice;
-  });
+liveSubscription.value = await AssetPrice.liveQuery(props.symbol, assetPrice => {
+  price.value = assetPrice;
 });
 
 onUnmounted(async () => {
   if (liveSubscription) {
-    await liveSubscription.unsubscribe();
+    await liveSubscription.value.unsubscribe();
   }
 });